+++ sketch-extensions-compose-resources/src/commonMain/kotlin/com/github/panpf/sketch/state/SaveCellularTrafficExtensions.composeResources.kt
@@ -1,0 +1,14 @@
+package com.github.panpf.sketch.state
+
+import androidx.compose.runtime.Composable
+import org.jetbrains.compose.resources.DrawableResource
+
+/**
+ * Set the error image when the save cellular traffic
+ */
+@Composable
+fun ErrorStateImage.Builder.saveCellularTrafficError(
+    resource: DrawableResource
+): ErrorStateImage.Builder = apply {
+    addState(SaveCellularTrafficCondition, resource)
+}