+++ sketch-extensions-compose-resources/src/commonTest/kotlin/com/github/panpf/sketch/extensions/compose/resources/test/SaveCellularTrafficExtensionsComposeResourcesTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.extensions.compose.resources.test
+
+class SaveCellularTrafficExtensionsComposeResourcesTest {
+    // TODO test
+}