+++ sketch-extensions-compose/src/commonMain/kotlin/com/github/panpf/sketch/state/SaveCellularTrafficExtensions.compose.kt
@@ -1,0 +1,13 @@
+package com.github.panpf.sketch.state
+
+import androidx.compose.ui.graphics.Color
+
+
+/**
+ * Set the error image when the save cellular traffic
+ */
+fun ErrorStateImage.Builder.saveCellularTrafficError(
+    color: Color
+): ErrorStateImage.Builder = apply {
+    addState(SaveCellularTrafficCondition, ColorPainterStateImage(color))
+}