+++ sketch-extensions-compose/src/commonTest/kotlin/com/github/panpf/sketch/extensions/compose/test/state/SaveCellularTrafficExtensionsAndroidTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.extensions.compose.test.state
+
+class SaveCellularTrafficExtensionsAndroidTest {
+    // TODO test
+}